+++ talpid-routing/src/unix/macos/data.rs
@@ -179,8 +179,7 @@ impl RouteMessage {
         self
     }
 
-    pub fn set_gateway_addr(mut self, addr: IpAddr) -> Self {
-        let gateway: SocketAddr = (addr, 0).into();
+    pub fn set_gateway_addr(mut self, gateway: impl Into<SockaddrStorage>) -> Self {
         self.insert_sockaddr(RouteSocketAddress::Gateway(Some(gateway.into())));
         self.route_flags |= RouteFlag::RTF_GATEWAY;